requirements.txt
# =============================================================================
# Odoo SaaS Platform Dependencies
# =============================================================================

# Core Flask Framework
Flask==2.3.3
Werkzeug==2.3.7

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
alembic==1.12.1
Flask-Migrate==4.0.5

# Authentication & Security
Flask-JWT-Extended==4.5.3
PyJWT==2.8.0
bcrypt==4.0.1
cryptography==41.0.7

# API & Validation
marshmallow==3.20.1
Flask-CORS==4.0.0
Flask-Limiter==3.5.0

# Redis & Caching
redis==4.6.0
Flask-Caching==2.1.0

# Background Jobs
rq==1.15.1
rq-dashboard==0.6.3

# HTTP Client for webhooks
requests==2.31.0
urllib3==2.0.7

# Environment & Configuration
python-dotenv==1.0.0
click==8.1.7

# Cloud Storage (AWS S3)
boto3==1.29.7
botocore==1.32.7

# Billing Integrations
stripe==5.7.0
# paddle-billing-python  # Install via pip install git+https://github.com/PaddleHQ/paddle-python-sdk.git

# Monitoring & Observability
prometheus-client==0.18.0
structlog==23.2.0

# Email
Flask-Mail==0.9.1

# Date/Time utilities
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==19.12.0

# Code Quality
black==23.9.1
flake8==6.1.0
isort==5.12.0
mypy==1.6.1

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Deployment
gunicorn==21.2.0
gevent==23.9.1

# Kubernetes client (optional)
kubernetes==27.2.0

# Development utilities
ipython==8.17.2
ipdb==0.13.13

# Data serialization
PyYAML==6.0.1
toml==0.10.2

# File handling
Pillow==10.0.1

# Math & Calculations
numpy==1.24.3
pandas==2.0.3

# UUID generation
uuid==1.30

# Process management
supervisor==4.2.5

# Health checks
healthcheck==1.3.3

# Configuration management
pydantic==2.4.2
pydantic[email]==2.4.2

# Time zone handling
zoneinfo==0.2.1; python_version < '3.9'

# Async support
aiohttp==3.8.6
asyncio==3.4.3

# XML/HTML parsing (for webhook processing)
lxml==4.9.3
beautifulsoup4==4.12.2

# Regular expressions
regex==2023.10.3